Updates permissions

+++ src/models/products.model.ts
@@ -53,14 +53,18 @@ export const getProductsByCategoryModel = (
   )
 }
 
-export const createProductModel = (req: Request, next: any): any => {
+export const createProductModel = (req: any, next: any): any => {
   const { name, price, categoryId, image, enabled, stock } = req.body
+  const role = req.user.role
+  if (role === 'client') {
+    return next({ message: 'You are not authorized to do this' }, null)
+  }
   pool.query(
     `INSERT INTO products (name, price, image, enabled, stock) VALUES ($1, $2, $3, $4, $5) RETURNING id`,
     [name, price, image ?? null, enabled ?? true, stock],
     (error, results) => {
       if (error) {
-        return next(error, null)
+        return next({ message: 'Error creating product' }, null)
       }
       const productId = results.rows[0].id
       pool.query(
@@ -68,7 +72,7 @@ export const createProductModel = (req: Request, next: any): any => {
         [productId, categoryId],
         (error, _results) => {
           if (error) {
-            return next(error, null)
+            return next({ message: 'Error creating product' }, null)
           }
           return next(null, {
             message: `New product added with ID: ${productId}`,
@@ -79,10 +83,13 @@ export const createProductModel = (req: Request, next: any): any => {
   )
 }
 
-export const updateProductModel = (req: Request, next: any): any => {
+export const updateProductModel = (req: any, next: any): any => {
   const id = parseInt(req.params.id)
   const { name, price, image, enabled, stock } = req.body
-
+  const role = req.user.role
+  if (role === 'client') {
+    return next({ message: 'You are not authorized to do this' }, null)
+  }
   getProductByIdModel(id, (error: Error, results: any) => {
     if (error) {
       return next(error, null)
@@ -110,7 +117,7 @@ export const likeProductModel = (req: any, next: any): any => {
     `INSERT INTO likes_users_products (id_user, id_product, likes) VALUES ($1, $2, $3) 
       ON CONFLICT (id_user, id_product) DO UPDATE SET id_user = $1, id_product = $2, likes = $3`,
     [id_user, id_product, likes],
-    (error, results) => {
+    (error, _results) => {
       if (error) {
         return next(error, null)
       }
@@ -121,20 +128,25 @@ export const likeProductModel = (req: any, next: any): any => {
   )
 }
 
-export const deleteProductModel = (id: number, next: any): any => {
+export const deleteProductModel = (req: any, next: any): any => {
+  const id = parseInt(req.params.id)
+  const role = req.user.role
+  if (role === 'client') {
+    return next({ message: 'You are not authorized to do this' }, null)
+  }
   pool.query(
     'DELETE FROM products_categories WHERE id_product = $1',
     [id],
     (error, _results) => {
       if (error) {
-        return next(error, null)
+        return next({ message: 'Error deleting product' }, null)
       }
       pool.query(
         'DELETE FROM products WHERE id = $1',
         [id],
         (error, _results) => {
           if (error) {
-            return next(error, null)
+            return next({ message: 'Error deleting product' }, null)
           }
           return next(null, { message: `Product deleted with ID: ${id}` })
         },
